serpapi-job-finder-cli/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "job-finder"
version = "0.1.0"
description = "CLI for searching Google Jobs via SerpApi"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
exclude = [".venv", "__pycache__", ".git", ".pytest_cache"]

[tool.ruff.lint.isort]
known-first-party = ["job_finder"]
